Building 3 and 2 addons (2 should be finished)

+++ Polimi On Fire/Assets/TurnTheGameOn/Arrow WayPointer/Scripts/Waypoint.cs
@@ -4,7 +4,7 @@
 	public class Waypoint : MonoBehaviour
 	{
 
-		public int radius;
+		public float radius;
 		[HideInInspector] public WaypointController waypointController;
 		[HideInInspector] public int waypointNumber;
